requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
scikit-learn
requests
deep-translator
python-dateutil
